All the functions for calculating the bolometric luminosity for AGNs are now in gne_Lagn.py
